+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 # ChatGPT Changelog
 ## [Unreleased]
+## [2.1.4-222.223]
+### Added
++ Support conversation model selection
++ Support GPT-3.5-Turbo context
++ Support system roles of GPT-3.5-Turbo
++ Support querying the usage limit of the account
++ Support for creating API Key
+
+## [2.1.3-222.223]
+### Added
++ Support for adding custom prefixes
+
+## [2.1.2-222.223]
+### Added
++ Rollback the Online ChatGPT.
++ Support custom display order of Content in tool windows
+
+### Changed
++ UI description information optimization
+
+
 ## [2.1.1-222.223]
 ### Changed
 + Fix some issues